+++ Plugin.cs
@@ -42,6 +42,12 @@ namespace DiscordBot
         Commands,
         DeathFeed
     }
+
+    public enum Event
+    {
+        Startup, Shutdown, Save, Raid, Login, Logout, Death
+    }
+    
     public enum Channel { Chat, Commands }
     public enum ChatDisplay { Player, Bot }
     
@@ -64,6 +70,21 @@ namespace DiscordBot
         
         public static DiscordBotPlugin m_instance = null!;
         
+        
+        // Extra Webhook URL
+        private static ConfigEntry<string> m_startWorldHook = null!;
+        
+        private static ConfigEntry<string> m_saveWebhook = null!;
+
+        private static ConfigEntry<string> m_shutdownWebhook = null!;
+
+        private static ConfigEntry<string> m_loginWebhook = null!;
+
+        private static ConfigEntry<string> m_logoutWebhook = null!;
+
+        private static ConfigEntry<string> m_eventWebhook = null!;
+        //
+        
         private static ConfigEntry<string> m_notificationWebhookURL = null!;
         private static ConfigEntry<Toggle> m_serverStartNotice = null!;
         private static ConfigEntry<Toggle> m_serverStopNotice = null!;
@@ -71,6 +92,7 @@ namespace DiscordBot
         private static ConfigEntry<Toggle> m_deathNotice = null!;
         private static ConfigEntry<Toggle> m_loginNotice = null!;
         private static ConfigEntry<Toggle> m_logoutNotice = null!;
+        private static ConfigEntry<Toggle> m_eventNotice = null!;
 
         private static ConfigEntry<string> m_chatWebhookURL = null!;
         private static ConfigEntry<string> m_chatChannelID = null!;
@@ -106,6 +128,7 @@ namespace DiscordBot
         public static bool ShowOnDeath => m_deathNotice.Value is Toggle.On;
         public static bool ShowOnLogin => m_loginNotice.Value is Toggle.On;
         public static bool ShowOnLogout => m_logoutNotice.Value is Toggle.On;
+        public static bool ShowEvent => m_eventNotice.Value is Toggle.On;
         public static ChatDisplay ChatType => m_chatType.Value;
         public static string DiscordAdmins => m_discordAdmins.Value;
         public static void SetDiscordAdmins(string value) => m_discordAdmins.Value = value;
@@ -124,7 +147,6 @@ namespace DiscordBot
         public static Resolution ScreenshotResolution => resolutions[m_screenshotResolution.Value];
         public static Resolution GifResolution => resolutions[m_gifResolution.Value];
         public static int ScreenshotDepth => m_screenshotDepth.Value;
-        
         public static KeyCode SelfieKey => m_selfieKey.Value;
         
         public static void LogWarning(string message) => DiscordBotLogger.LogWarning(message);
@@ -132,6 +154,14 @@ namespace DiscordBot
 
         private static readonly Dictionary<string, Resolution> resolutions = new();
 
+        public static List<string> OnWorldStartHooks => m_startWorldHook.Value.IsNullOrWhiteSpace()
+            ? new List<string>()
+            : new StringListConfig(m_startWorldHook.Value).list;
+        public static List<string> OnWorldSaveHooks => m_saveWebhook.Value.IsNullOrWhiteSpace() ? new List<string>() : new StringListConfig(m_saveWebhook.Value).list;
+        public static List<string> OnWorldShutdownHooks => m_shutdownWebhook.Value.IsNullOrWhiteSpace() ?  new List<string>() : new StringListConfig(m_shutdownWebhook.Value).list;
+        public static List<string> OnLoginHooks => m_loginWebhook.Value.IsNullOrWhiteSpace() ?  new List<string>() : new StringListConfig(m_loginWebhook.Value).list;
+        public static List<string> OnLogoutHooks => m_logoutWebhook.Value.IsNullOrWhiteSpace() ? new List<string>() : new StringListConfig(m_logoutWebhook.Value).list;
+        public static List<string> OnEventHooks => m_eventWebhook.Value.IsNullOrWhiteSpace() ? new List<string>() : new StringListConfig(m_eventWebhook.Value).list;
         
         // TODO : Figure out to make sure connecting peer is connecting to the right server
 
@@ -150,6 +180,7 @@ namespace DiscordBot
             m_serverSaveNotice = config("2 - Notifications", "Saving", Toggle.On, "If on, bot will send message when server is saving");
             m_loginNotice = config("2 - Notifications", "Login", Toggle.On, "If on, bot will send message when player logs in");
             m_logoutNotice = config("2 - Notifications", "Logout", Toggle.On, "If on, bot will send message when player logs out");
+            m_eventNotice = config("2 - Notifications", "Random Events", Toggle.On, "If on, bot will send message when random event starts");
 
             m_chatWebhookURL = config("3 - Chat", "Webhook URL", "", "Set discord webhook to display chat messages");
             m_chatChannelID = config("3 - Chat", "Channel ID", "", "Set channel ID to monitor for messages");
@@ -209,6 +240,15 @@ namespace DiscordBot
             m_gifResolution.SettingChanged += (_, _) => Recorder.instance?.OnGifResolutionChange();
 
             m_selfieKey = config("1 - General", "Selfie", KeyCode.None, "Hotkey to take selfie and send to discord", false);
+
+            m_startWorldHook = config("7 - Webhooks", "On World Start", "",
+                new ConfigDescription("If empty, will use default notification webhook", null,
+                    new ConfigurationManagerAttributes() { CustomDrawer = StringListConfig.Draw }));
+            m_saveWebhook = config("7 - Webhooks", "On World Save", "", new ConfigDescription("If empty, will use default notification webhook", null, new ConfigurationManagerAttributes(){ CustomDrawer = StringListConfig.Draw}));
+            m_shutdownWebhook = config("7 - Webhooks", "On World Shutdown", "", new ConfigDescription("If empty, will use default notification webhook", null, new ConfigurationManagerAttributes(){ CustomDrawer = StringListConfig.Draw}));
+            m_loginWebhook = config("7 - Webhooks", "On Login",  "", new ConfigDescription("If empty, will use default notification webhook", null, new ConfigurationManagerAttributes(){ CustomDrawer = StringListConfig.Draw}));
+            m_logoutWebhook = config("7 - Webhooks", "On Logout", "", new ConfigDescription("If empty, will use default notification webhook", null, new ConfigurationManagerAttributes(){ CustomDrawer = StringListConfig.Draw}));
+            m_eventWebhook = config("7 - Webhooks", "On Event", "", new ConfigDescription("If empty, will use default notification webhook", null, new ConfigurationManagerAttributes(){ CustomDrawer = StringListConfig.Draw}));
             
             DiscordCommands.Setup();
             DeathQuips.Setup();